move emip3 and tx_builder_constants tests to test module
