+++ rust/src/tests/builders/mod.rs
@@ -5,3 +5,4 @@ mod voting_proposal_builder;
 mod certificates_builder;
 mod mint_builder;
 mod tx_inputs_builder;
+mod tx_builder_constans;
